+++ src/components/Search/types.ts
@@ -1,11 +1,12 @@
 import { ReactElement } from 'react';
+import TShape from '../../types/shapes';
 import { TAlignment } from '../Field/types';
 
 export interface IProps {
   label?: string;
   labelWidth?: string;
   labelAlign?: TAlignment;
-  shape?: 'round' | 'square';
+  shape?: TShape;
   background?: string;
   maxLength?: number;
   placeholder?: string;