+++ src/index.tsx
@@ -11,6 +11,7 @@ import Image from './components/Image';
 import Slider from './components/Slider';
 import Checkbox from './components/Checkbox';
 import Radio from './components/Radio';
+import Stepper from './components/Stepper';
 
 export { default as Button } from './components/Button';
 export { default as Icon } from './components/Icons';
@@ -25,6 +26,7 @@ export { default as Image } from './components/Image';
 export { default as Slider } from './components/Slider';
 export { default as Checkbox } from './components/Checkbox';
 export { default as Radio } from './components/Radio';
+export { default as Stepper } from './components/Stepper';
 
 const Vant = {
   Button,
@@ -39,7 +41,8 @@ const Vant = {
   Image,
   Slider,
   Checkbox,
-  Radio
+  Radio,
+  Stepper
 };
 
 export default Vant;